+++ packages/orq-rc/RELEASES.md
@@ -12599,3 +12599,13 @@ Based on:
 - [typescript v4.16.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.15 - packages/orq-rc
+
+## 2026-09-23 15:39:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.798.0 (2.938.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.16 - packages/orq-rc